refactor(exploit-intelligence): move business logic from endpoints to runner module
Move background processing functions (run_analysis, poll_for_result,
poll_for_product_result, recover_and_resume) out of the endpoints layer
into a new runner/ module. Endpoints now contain only HTTP handlers.
