+++ modules/fundamental/src/exploit_intelligence/endpoints/mod.rs
@@ -8,84 +8,19 @@ use crate::{
             AnalyzeRequest, AnalyzeResponse, ExploitIntelligenceJobDetails,
             ExploitIntelligenceJobSummary, JobStatus,
         },
+        runner::{analysis::run_analysis, recovery::recover_and_resume},
         service::ExploitIntelligenceService,
     },
 };
 use actix_web::{HttpResponse, Responder, get, post, web};
-use futures_util::TryStreamExt;
-use sea_orm::{AccessMode, ColumnTrait, EntityTrait, QueryFilter, TransactionTrait};
-use std::str::FromStr;
+use sea_orm::{ColumnTrait, EntityTrait, QueryFilter, TransactionTrait};
 use trustify_auth::{CreateExploitIntelligence, ReadExploitIntelligence, authorizer::Require};
 use trustify_common::{
     db,
     model::{Paginated, PaginatedResults},
 };
-use backon::{ExponentialBuilder, ConstantBuilder, Retryable};
-use std::cell::Cell;
-use trustify_entity::labels::Labels;
-use trustify_module_ingestor::service::{Cache, Format, IngestorService};
-use trustify_module_storage::service::StorageBackend;
+use trustify_module_ingestor::service::IngestorService;
 use uuid::Uuid;
-
-/// Error classification for EI HTTP requests.
-enum EiRequestError {
-    /// Retryable failure (network timeout, 5xx, etc.)
-    Transient(String),
-    /// Fatal failure (auth error, 4xx, parse error)
-    Permanent(anyhow::Error),
-}
-
-impl EiRequestError {
-    fn is_transient(&self) -> bool {
-        matches!(self, Self::Transient(_))
-    }
-
-    fn into_anyhow(self) -> anyhow::Error {
-        match self {
-            Self::Transient(msg) => anyhow::anyhow!(msg),
-            Self::Permanent(err) => err,
-        }
-    }
-}
-
-impl std::fmt::Display for EiRequestError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Transient(msg) => write!(f, "transient: {msg}"),
-            Self::Permanent(err) => write!(f, "permanent: {err}"),
-        }
-    }
-}
-
-/// Classify an HTTP response or reqwest error into an EiRequestError.
-fn classify_response_error(
-    result: Result<reqwest::Response, reqwest::Error>,
-) -> Result<reqwest::Response, EiRequestError> {
-    match result {
-        Ok(resp) => {
-            let status = resp.status();
-            if status.is_success() {
-                Ok(resp)
-            } else {
-                match status.as_u16() {
-                    408 | 429 | 502 | 503 | 504 => Err(EiRequestError::Transient(format!(
-                        "transient HTTP status {status}"
-                    ))),
-                    _ => Err(EiRequestError::Permanent(anyhow::anyhow!("HTTP {status}"))),
-                }
-            }
-        }
-        Err(err) => {
-            if err.is_connect() || err.is_timeout() || err.is_request() {
-                Err(EiRequestError::Transient(format!("network error: {err}")))
-            } else {
-                Err(EiRequestError::Permanent(anyhow::anyhow!(
-                    "network error: {err}"
-                )))
-            }
-        }
-    }
-}
 
 /// Register exploit intelligence endpoints and spawn recovery for orphaned jobs.
 pub fn configure(
@@ -112,117 +47,6 @@ pub fn configure(
         .service(analyze)
         .service(list_jobs)
         .service(get_job);
-}
-
-/// Recover orphaned EI jobs after a server restart.
-///
-/// Pending jobs (never submitted) are marked as Failed.  Running jobs
-/// (scan_id present) get their poll loops re-spawned so the EI backend
-/// result is picked up transparently.
-async fn recover_and_resume(
-    ei_service: &ExploitIntelligenceService,
-    ingestor: &IngestorService,
-    db_rw: &db::ReadWrite,
-) {
-    let result = match ei_service.recover_orphaned_jobs(db_rw).await {
-        Ok(r) => r,
-        Err(e) => {
-            tracing::error!(error = %e, "failed to recover orphaned exploit intelligence jobs");
-            return;
-        }
-    };
-
-    let runtime = match ei_service.runtime().cloned() {
-        Some(rt) => rt,
-        None => return,
-    };
-    let config = runtime.config;
-    let client = runtime.client;
-
-    for job in result.running_jobs {
-        let job_id = job.id;
-        let vulnerability_id = job.vulnerability_id.clone();
-        let ei_svc = ei_service.clone();
-        let ingestor_svc = ingestor.clone();
-        let db_rw_clone = db_rw.clone();
-        let config_clone = config.clone();
-        let client_clone = client.clone();
-
-        if let Some(ref product_id) = job.product_id {
-            // Multi-component SPDX flow — resume product-level polling.
-            let product_id = product_id.clone();
-            actix_web::rt::spawn(async move {
-                tracing::info!(job_id = %job_id, product_id = %product_id, "resuming orphaned SPDX product polling");
-                if let Err(e) = poll_for_product_result(
-                    &ei_svc,
-                    &ingestor_svc,
-                    &db_rw_clone,
-                    &config_clone,
-                    &client_clone,
-                    job_id,
-                    &vulnerability_id,
-                    &product_id,
-                )
-                .await
-                {
-                    tracing::error!(job_id = %job_id, error = %e, "failed to resume orphaned SPDX product polling");
-                    let _ = ei_svc
-                        .update_job_failed(job_id, &e.to_string(), &db_rw_clone)
-                        .await;
-                }
-            });
-        } else {
-            // Single-component CycloneDX flow — look up scan_id from the component.
-            let components = match ei_service.fetch_components(job_id, db_rw).await {
-                Ok(c) => c,
-                Err(e) => {
-                    tracing::error!(job_id = %job_id, error = %e, "failed to fetch components during recovery");
-                    let _ = ei_service
-                        .update_job_failed(job_id, &format!("recovery error: {e}"), db_rw)
-                        .await;
-                    continue;
-                }
-            };
-
-            let first_component = components.first();
-            let scan_id = first_component.and_then(|c| c.scan_id.as_deref());
-            let component_id = first_component.map(|c| c.id);
-
-            if let (Some(scan_id), Some(component_id)) = (scan_id, component_id) {
-                let scan_id = scan_id.to_string();
-                actix_web::rt::spawn(async move {
-                    tracing::info!(job_id = %job_id, scan_id = %scan_id, "resuming orphaned EI analysis");
-                    if let Err(e) = poll_for_result(
-                        &ei_svc,
-                        &ingestor_svc,
-                        &db_rw_clone,
-                        &config_clone,
-                        &client_clone,
-                        job_id,
-                        &vulnerability_id,
-                        &scan_id,
-                        component_id,
-                    )
-                    .await
-                    {
-                        tracing::error!(job_id = %job_id, error = %e, "failed to resume orphaned EI analysis");
-                        let _ = ei_svc
-                            .update_job_failed(job_id, &e.to_string(), &db_rw_clone)
-                            .await;
-                    }
-                });
-            } else {
-                tracing::error!(job_id = %job_id, "running job has no component with scan_id, marking failed");
-                let _ = ei_service
-                    .update_job_failed(
-                        job_id,
-                        "Running job missing component scan_id during recovery",
-                        db_rw,
-                    )
-                    .await;
-            }
-        }
-    }
 }
 
 /// Query parameters for filtering the job list.
@@ -384,798 +208,5 @@ pub async fn get_job(
     match job {
         Some(details) => Ok(HttpResponse::Ok().json(details)),
         None => Ok(HttpResponse::NotFound().finish()),
-    }
-}
-
-/// Background task: submit SBOM to EI client, poll for results, ingest VEX.
-async fn run_analysis(
-    ei_service: &ExploitIntelligenceService,
-    ingestor: &IngestorService,
-    db_rw: &db::ReadWrite,
-    config: &crate::exploit_intelligence::service::ExploitIntelligenceConfig,
-    client: &reqwest::Client,
-    job_id: Uuid,
-    sbom_id: Uuid,
-    vulnerability_id: &str,
-) -> Result<(), anyhow::Error> {
-    // Retrieve the raw SBOM document from storage (read-only – no mutations needed).
-    let tx = db_rw
-        .begin_with_config(None, Some(AccessMode::ReadOnly))
-        .await?;
-    let sbom_model = trustify_entity::sbom::Entity::find()
-        .filter(trustify_entity::sbom::Column::SbomId.eq(sbom_id))
-        .one(&tx)
-        .await?
-        .ok_or_else(|| anyhow::anyhow!("SBOM not found: {sbom_id}"))?;
-
-    let source_doc =
-        trustify_entity::source_document::Entity::find_by_id(sbom_model.source_document_id)
-            .one(&tx)
-            .await?
-            .ok_or_else(|| anyhow::anyhow!("source document not found for SBOM: {sbom_id}"))?;
-
-    // Build storage key from source document SHA256.
-    // The DB stores the raw hex digest without algorithm prefix, but Id::from_str expects "sha256:…".
-    let sha256_with_prefix = format!("sha256:{}", source_doc.sha256);
-    let storage_key = trustify_common::id::Id::from_str(&sha256_with_prefix)
-        .map_err(|e| anyhow::anyhow!("invalid sha256: {e}"))?;
-    let storage_key: trustify_module_storage::service::StorageKey = storage_key
-        .try_into()
-        .map_err(|e: trustify_module_storage::service::StorageKeyError| {
-            anyhow::anyhow!("storage key error: {e}")
-        })?;
-
-    let stream = ingestor
-        .storage()
-        .retrieve(storage_key)
-        .await
-        .map_err(|e| anyhow::anyhow!("storage retrieve error: {e}"))?
-        .ok_or_else(|| anyhow::anyhow!("SBOM document not found in storage"))?;
-
-    // Collect the stream into bytes.
-    let sbom_bytes: Vec<u8> = stream
-        .try_fold(Vec::new(), |mut acc, chunk| async move {
-            acc.extend_from_slice(&chunk);
-            Ok(acc)
-        })
-        .await
-        .map_err(|e| anyhow::anyhow!("failed to read SBOM stream: {e}"))?;
-
-    drop(tx);
-
-    // Detect SBOM format from labels.
-    let sbom_format = match sbom_model.labels.0.get("type") {
-        Some(t) => t.as_str(),
-        None => {
-            tracing::warn!(
-                sbom_id = %sbom_id,
-                "SBOM has no 'type' label, defaulting to CycloneDX format"
-            );
-            "cyclonedx"
-        }
-    };
-
-    let is_spdx = sbom_format == "spdx";
-
-    if is_spdx {
-        // SPDX multi-component flow: upload to /api/v1/products/upload-spdx
-        let upload_url = format!("{}/api/v1/products/upload-spdx", config.url);
-
-        let response = (|| {
-            let url = upload_url.clone();
-            let sbom = sbom_bytes.clone();
-            let vuln_id = vulnerability_id.to_string();
-            let cli = client.clone();
-            let tp = config.token_provider.clone();
-            async move {
-                let form = reqwest::multipart::Form::new()
-                    .text("cveId", vuln_id)
-                    .part(
-                        "file",
-                        reqwest::multipart::Part::bytes(sbom)
-                            .file_name("sbom.json")
-                            .mime_str("application/json")
-                            .map_err(|e| {
-                                EiRequestError::Permanent(anyhow::anyhow!("mime error: {e}"))
-                            })?,
-                    );
-
-                let mut request = cli.post(&url).multipart(form);
-
-                if let Some(ref provider) = tp {
-                    let token = provider.get_token().await.map_err(|e| {
-                        EiRequestError::Permanent(anyhow::anyhow!(
-                            "failed to obtain auth token: {e}"
-                        ))
-                    })?;
-                    request = request.bearer_auth(&token);
-                }
-
-                classify_response_error(request.send().await)
-            }
-        })
-        .retry(ExponentialBuilder::default().with_min_delay(std::time::Duration::from_secs(1)).with_max_times(3))
-        .when(|e| e.is_transient())
-        .notify(|e, dur| tracing::warn!(?dur, %e, "retrying EI upload"))
-        .await
-        .map_err(EiRequestError::into_anyhow)?;
-
-        let spdx_upload: crate::exploit_intelligence::ei_client::SpdxUploadResponse =
-            response.json().await?;
-        let product_id = spdx_upload.product_id;
-
-        // Update job to running with the product ID.
-        ei_service
-            .update_job_running_with_product(job_id, &product_id, db_rw)
-            .await
-            .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-
-        tracing::info!(
-            job_id = %job_id,
-            product_id = %product_id,
-            "SPDX product uploaded to EI, starting product-level polling"
-        );
-
-        // Hand off to the product-level poll loop.
-        poll_for_product_result(
-            ei_service,
-            ingestor,
-            db_rw,
-            config,
-            client,
-            job_id,
-            vulnerability_id,
-            &product_id,
-        )
-        .await
-    } else {
-        // CycloneDX single-component flow.
-        // Create the component record before uploading so it is tracked.
-        let component_ref = sbom_model.document_id.clone().unwrap_or_else(|| sbom_model.node_id.clone());
-        let component_id = ei_service
-            .create_component(job_id, &component_ref, db_rw)
-            .await
-            .map_err(|e| anyhow::anyhow!("failed to create component: {e}"))?;
-
-        let upload_url = format!("{}/api/v1/products/upload-cyclonedx", config.url);
-
-        let response = (|| {
-            let url = upload_url.clone();
-            let sbom = sbom_bytes.clone();
-            let vuln_id = vulnerability_id.to_string();
-            let cli = client.clone();
-            let tp = config.token_provider.clone();
-            async move {
-                let form = reqwest::multipart::Form::new()
-                    .text("cveId", vuln_id)
-                    .part(
-                        "file",
-                        reqwest::multipart::Part::bytes(sbom)
-                            .file_name("sbom.json")
-                            .mime_str("application/json")
-                            .map_err(|e| {
-                                EiRequestError::Permanent(anyhow::anyhow!("mime error: {e}"))
-                            })?,
-                    );
-
-                let mut request = cli.post(&url).multipart(form);
-
-                if let Some(ref provider) = tp {
-                    let token = provider.get_token().await.map_err(|e| {
-                        EiRequestError::Permanent(anyhow::anyhow!(
-                            "failed to obtain auth token: {e}"
-                        ))
-                    })?;
-                    request = request.bearer_auth(&token);
-                }
-
-                classify_response_error(request.send().await)
-            }
-        })
-        .retry(ExponentialBuilder::default().with_min_delay(std::time::Duration::from_secs(1)).with_max_times(3))
-        .when(|e| e.is_transient())
-        .notify(|e, dur| tracing::warn!(?dur, %e, "retrying EI upload"))
-        .await
-        .map_err(EiRequestError::into_anyhow)?;
-
-        let upload: crate::exploit_intelligence::ei_client::UploadResponse = response.json().await?;
-        let scan_id = upload.report_request_id.report_id;
-
-        // Update component to running with scan ID.
-        ei_service
-            .update_component_running(component_id, &scan_id, db_rw)
-            .await
-            .map_err(|e| anyhow::anyhow!("failed to update component: {e}"))?;
-
-        // Hand off to the poll loop.
-        poll_for_result(
-            ei_service,
-            ingestor,
-            db_rw,
-            config,
-            client,
-            job_id,
-            vulnerability_id,
-            &scan_id,
-            component_id,
-        )
-        .await
-    }
-}
-
-/// Poll the EI service for an in-flight single-component analysis and ingest the result.
-///
-/// This is the second half of the analysis lifecycle: it assumes the SBOM has
-/// already been uploaded and a `scan_id` obtained.  It is called both from
-/// `run_analysis` (normal path) and from the startup recovery path for
-/// orphaned Running jobs.
-///
-/// Updates both the component record (with finding/advisory) and the parent
-/// job record (status only) upon completion.
-pub(crate) async fn poll_for_result(
-    ei_service: &ExploitIntelligenceService,
-    ingestor: &IngestorService,
-    db_rw: &db::ReadWrite,
-    config: &crate::exploit_intelligence::service::ExploitIntelligenceConfig,
-    client: &reqwest::Client,
-    job_id: Uuid,
-    _vulnerability_id: &str,
-    scan_id: &str,
-    component_id: Uuid,
-) -> Result<(), anyhow::Error> {
-    let poll_url = format!(
-            "{}/api/v1/reports/by-scan-id/{}",
-            config.url,
-            urlencoding::encode(scan_id)
-        );
-    let poll_secs = config.poll_interval_secs.max(5);
-    let consecutive_failures = Cell::new(0u32);
-
-    let result = tokio::time::timeout(
-        std::time::Duration::from_secs(config.max_poll_duration_secs),
-        (|| {
-            let url = poll_url.clone();
-            let cli = client.clone();
-            let tp = config.token_provider.clone();
-            async move {
-                let mut poll_request = cli.get(&url);
-                if let Some(ref provider) = tp {
-                    let token = provider.get_token().await.map_err(|e| {
-                        EiRequestError::Permanent(anyhow::anyhow!(
-                            "failed to obtain auth token: {e}"
-                        ))
-                    })?;
-                    poll_request = poll_request.bearer_auth(&token);
-                }
-
-                let resp = classify_response_error(poll_request.send().await)?;
-
-                let poll: crate::exploit_intelligence::ei_client::PollResponse =
-                    resp.json().await.map_err(|e| {
-                        EiRequestError::Permanent(anyhow::anyhow!(
-                            "failed to parse EI poll response: {e}"
-                        ))
-                    })?;
-
-                match poll.status.as_str() {
-                    "completed" | "failed" | "expired" => Ok(poll),
-                    _ => {
-                        tracing::debug!(
-                            status = poll.status,
-                            "exploit intelligence analysis still in progress"
-                        );
-                        Err(EiRequestError::Transient("pending".to_string()))
-                    }
-                }
-            }
-        })
-        .retry(
-            ConstantBuilder::default()
-                .with_delay(std::time::Duration::from_secs(poll_secs))
-                .without_max_times(),
-        )
-        .when(|e| match e {
-            EiRequestError::Transient(reason) if reason == "pending" => {
-                consecutive_failures.set(0);
-                true
-            }
-            EiRequestError::Transient(_) => {
-                let n = consecutive_failures.get() + 1;
-                consecutive_failures.set(n);
-                n < 5
-            }
-            EiRequestError::Permanent(_) => false,
-        })
-        .notify(|e, dur| {
-            if !matches!(e, EiRequestError::Transient(r) if r == "pending") {
-                tracing::warn!(?dur, %e, "transient poll failure");
-            }
-        }),
-    )
-    .await;
-
-    // Convert timeout/retry result to our expected shape.
-    let poll = match result {
-        Ok(Ok(poll)) => poll,
-        Ok(Err(e)) => {
-            let err = e.into_anyhow();
-            let msg = err.to_string();
-            ei_service
-                .update_job_failed(job_id, &msg, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-            return Err(err);
-        }
-        Err(_elapsed) => {
-            let msg = format!("polling timed out after {}s", config.max_poll_duration_secs);
-            ei_service
-                .update_job_failed(job_id, &msg, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-            return Err(anyhow::anyhow!("{msg}"));
-        }
-    };
-
-    // Process the terminal poll response.
-    match poll.status.as_str() {
-        "completed" => {
-            // Extract finding from the analysis result.
-            use trustify_entity::exploit_intelligence_job::Finding;
-            let finding = poll
-                .report
-                .as_ref()
-                .and_then(|r| r.output.as_ref())
-                .and_then(|o| o.analysis.as_ref())
-                .and_then(|a| a.first())
-                .and_then(|e| e.justification.as_ref())
-                .and_then(|j| j.status.as_deref())
-                .map(|s| match s.to_ascii_lowercase().as_str() {
-                    "false" => Finding::NotVulnerable,
-                    "true" => Finding::Vulnerable,
-                    _ => Finding::Uncertain,
-                })
-                .unwrap_or(Finding::Uncertain);
-
-            // Check if VEX document is available.
-            let vex_json = poll
-                .report
-                .as_ref()
-                .and_then(|r| r.output.as_ref())
-                .and_then(|o| o.vex.as_ref());
-
-            let advisory_id = if let Some(vex) = vex_json {
-                if !vex.is_null() {
-                    // Ingest the VEX document as an advisory with source label.
-                    let vex_bytes = serde_json::to_vec(vex)?;
-                    let mut labels = Labels::default();
-                    labels
-                        .0
-                        .insert("source".to_string(), "exploit-intelligence".to_string());
-
-                    let tx = db_rw.begin().await?;
-                    let result = ingestor
-                        .ingest(&vex_bytes, Format::Advisory, labels, None, Cache::Skip, &tx)
-                        .await
-                        .map_err(|e| anyhow::anyhow!("VEX ingestion failed: {e}"))?;
-                    tx.commit().await?;
-
-                    tracing::info!(
-                        job_id = %job_id,
-                        advisory_id = %result.id,
-                        "ingested VEX advisory from exploit intelligence"
-                    );
-
-                    result
-                        .id
-                        .strip_prefix("urn:uuid:")
-                        .and_then(|id| Uuid::from_str(id).ok())
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-
-            // Update the component with the per-analysis results.
-            ei_service
-                .update_component_completed(component_id, finding, advisory_id, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update component: {e}"))?;
-
-            // Mark the parent job as completed (finding lives on the component).
-            ei_service
-                .update_job_completed(job_id, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-
-            tracing::info!(
-                job_id = %job_id,
-                finding = ?finding,
-                "exploit intelligence analysis completed"
-            );
-
-            Ok(())
-        }
-        "failed" | "expired" => {
-            let error_msg = poll
-                .report
-                .as_ref()
-                .and_then(|r| r.error.as_deref())
-                .unwrap_or("analysis failed")
-                .to_string();
-
-            ei_service
-                .update_job_failed(job_id, &error_msg, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-
-            Ok(())
-        }
-        _ => unreachable!("backon retry only returns terminal states"),
-    }
-}
-
-/// Poll the EI service for a multi-component product and process per-component results.
-///
-/// This is the SPDX counterpart of `poll_for_result`. It polls the product-level
-/// endpoint until all components are complete, then fetches each component's
-/// report individually.
-pub(crate) async fn poll_for_product_result(
-    ei_service: &ExploitIntelligenceService,
-    ingestor: &IngestorService,
-    db_rw: &db::ReadWrite,
-    config: &crate::exploit_intelligence::service::ExploitIntelligenceConfig,
-    client: &reqwest::Client,
-    job_id: Uuid,
-    _vulnerability_id: &str,
-    product_id: &str,
-) -> Result<(), anyhow::Error> {
-    let product_url = format!(
-            "{}/api/v1/products/{}",
-            config.url,
-            urlencoding::encode(product_id)
-        );
-    let poll_secs = config.poll_interval_secs.max(5);
-    let consecutive_failures = Cell::new(0u32);
-
-    let mut components_initialized = false;
-
-    // Maps EI scan_id -> our internal component_id (populated on first successful poll).
-    let mut scan_to_component: std::collections::HashMap<String, Uuid> =
-        std::collections::HashMap::new();
-
-    // Check whether component records already exist for this job (e.g. after a
-    // crash/restart that orphaned a running job). If they do, skip re-initialization
-    // to avoid duplicate component rows.
-    let existing_components = ei_service
-        .fetch_components(job_id, db_rw)
-        .await
-        .map_err(|e| anyhow::anyhow!("failed to fetch existing components: {e}"))?;
-    if !existing_components.is_empty() {
-        for c in &existing_components {
-            if let Some(ref scan_id) = c.scan_id {
-                scan_to_component.insert(scan_id.clone(), c.id);
-            }
-        }
-        components_initialized = true;
-        tracing::info!(
-            job_id = %job_id,
-            existing_components = existing_components.len(),
-            "resumed with existing component records, skipping re-initialization"
-        );
-    }
-
-    let result = tokio::time::timeout(
-        std::time::Duration::from_secs(config.max_poll_duration_secs),
-        (|| {
-            let url = product_url.clone();
-            let cli = client.clone();
-            let tp = config.token_provider.clone();
-            async move {
-                let mut poll_request = cli.get(&url);
-                if let Some(ref provider) = tp {
-                    let token = provider.get_token().await.map_err(|e| {
-                        EiRequestError::Permanent(anyhow::anyhow!(
-                            "failed to obtain auth token: {e}"
-                        ))
-                    })?;
-                    poll_request = poll_request.bearer_auth(&token);
-                }
-
-                let resp = classify_response_error(poll_request.send().await)?;
-
-                let product: crate::exploit_intelligence::ei_client::ProductStatusResponse =
-                    resp.json().await.map_err(|e| {
-                        EiRequestError::Permanent(anyhow::anyhow!(
-                            "failed to parse EI product response: {e}"
-                        ))
-                    })?;
-
-                match product.summary.product_state.as_str() {
-                    "completed" | "failed" => Ok(product),
-                    _ => {
-                        tracing::debug!(
-                            state = product.summary.product_state,
-                            "SPDX product analysis still in progress"
-                        );
-                        Err(EiRequestError::Transient("pending".to_string()))
-                    }
-                }
-            }
-        })
-        .retry(
-            ConstantBuilder::default()
-                .with_delay(std::time::Duration::from_secs(poll_secs))
-                .without_max_times(),
-        )
-        .when(|e| match e {
-            EiRequestError::Transient(reason) if reason == "pending" => {
-                consecutive_failures.set(0);
-                true
-            }
-            EiRequestError::Transient(_) => {
-                let n = consecutive_failures.get() + 1;
-                consecutive_failures.set(n);
-                n < 5
-            }
-            EiRequestError::Permanent(_) => false,
-        })
-        .notify(|e, dur| {
-            if !matches!(e, EiRequestError::Transient(r) if r == "pending") {
-                tracing::warn!(?dur, %e, "transient poll failure");
-            }
-        }),
-    )
-    .await;
-
-    // Convert timeout/retry result to our expected shape.
-    let product = match result {
-        Ok(Ok(product)) => product,
-        Ok(Err(e)) => {
-            let err = e.into_anyhow();
-            let msg = err.to_string();
-            ei_service
-                .update_job_failed(job_id, &msg, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-            return Err(err);
-        }
-        Err(_elapsed) => {
-            let msg = format!("polling timed out after {}s", config.max_poll_duration_secs);
-            ei_service
-                .update_job_failed(job_id, &msg, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-            return Err(anyhow::anyhow!("{msg}"));
-        }
-    };
-
-    // Process the terminal product response.
-    match product.summary.product_state.as_str() {
-        "completed" => {
-            tracing::info!(
-                job_id = %job_id,
-                product_id = %product_id,
-                "SPDX product analysis completed, fetching per-component reports"
-            );
-
-            // The EI product endpoint does not return per-component
-            // data (its `components` array is always empty). Instead,
-            // fetch the reports list filtered by product ID to discover
-            // each component's scan ID and justification.
-            let reports_url = format!(
-                "{}/api/v1/reports?productId={}",
-                config.url,
-                urlencoding::encode(product_id)
-            );
-
-            let mut reports_request = client.get(&reports_url);
-            if let Some(ref provider) = config.token_provider {
-                let token = provider
-                    .get_token()
-                    .await
-                    .map_err(|e| anyhow::anyhow!("failed to obtain auth token: {e}"))?;
-                reports_request = reports_request.bearer_auth(&token);
-            }
-
-            let reports_response = reports_request
-                .send()
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to fetch product reports: {e}"))?;
-
-            if !reports_response.status().is_success() {
-                let status = reports_response.status();
-                let body = reports_response.text().await.unwrap_or_default();
-                return Err(anyhow::anyhow!(
-                    "EI product reports request failed ({status}): {body}"
-                ));
-            }
-
-            let component_reports: Vec<crate::exploit_intelligence::ei_client::ProductReportEntry> =
-                reports_response.json().await?;
-
-            // Record submission failures as excluded components.
-            if !components_initialized {
-                for failure in &product.data.submission_failures {
-                    let image = failure.image.as_deref().unwrap_or("unknown");
-                    let error = failure.error.as_deref().unwrap_or("excluded by EI");
-
-                    ei_service
-                        .create_excluded_component(job_id, image, error, db_rw)
-                        .await
-                        .map_err(|e| anyhow::anyhow!("failed to create excluded component: {e}"))?;
-                }
-            }
-
-            tracing::info!(
-                job_id = %job_id,
-                product_id = %product_id,
-                reports = component_reports.len(),
-                excluded = product.data.submission_failures.len(),
-                "processing per-component reports for SPDX product"
-            );
-
-            use trustify_entity::exploit_intelligence_job::Finding;
-
-            for entry in &component_reports {
-                let scan_id = &entry.scan_id;
-                let image_name = &entry.image_name;
-
-                // Create a component record for this report.
-                let component_id = if !components_initialized {
-                    let id = ei_service
-                        .create_component(job_id, image_name, db_rw)
-                        .await
-                        .map_err(|e| anyhow::anyhow!("failed to create component: {e}"))?;
-                    ei_service
-                        .update_component_running(id, scan_id, db_rw)
-                        .await
-                        .map_err(|e| anyhow::anyhow!("failed to update component: {e}"))?;
-                    id
-                } else {
-                    match scan_to_component.get(scan_id) {
-                        Some(id) => *id,
-                        None => continue,
-                    }
-                };
-
-                // Handle non-completed reports (expired, failed on EI side).
-                if entry.state != "completed" {
-                    ei_service
-                        .update_component_failed(
-                            component_id,
-                            &format!("EI report state: {}", entry.state),
-                            db_rw,
-                        )
-                        .await
-                        .map_err(|e| anyhow::anyhow!("failed to update component: {e}"))?;
-                    continue;
-                }
-
-                // Extract finding from the justification in the reports list.
-                let finding = entry
-                    .vulns
-                    .first()
-                    .and_then(|v| v.justification.as_ref())
-                    .and_then(|j| j.status.as_deref())
-                    .map(|s| match s.to_ascii_lowercase().as_str() {
-                        "false" => Finding::NotVulnerable,
-                        "true" => Finding::Vulnerable,
-                        _ => Finding::Uncertain,
-                    })
-                    .unwrap_or(Finding::Uncertain);
-
-                // Fetch the full by-scan-id report for VEX ingestion.
-                let full_report_url = format!(
-                    "{}/api/v1/reports/by-scan-id/{}",
-                    config.url,
-                    urlencoding::encode(scan_id)
-                );
-
-                let mut full_request = client.get(&full_report_url);
-                if let Some(ref provider) = config.token_provider {
-                    let token = provider
-                        .get_token()
-                        .await
-                        .map_err(|e| anyhow::anyhow!("failed to obtain auth token: {e}"))?;
-                    full_request = full_request.bearer_auth(&token);
-                }
-
-                let advisory_id = match full_request.send().await {
-                    Ok(resp) if resp.status().is_success() => {
-                        let poll_resp: crate::exploit_intelligence::ei_client::PollResponse =
-                            resp.json().await?;
-
-                        let vex_json = poll_resp
-                            .report
-                            .as_ref()
-                            .and_then(|r| r.output.as_ref())
-                            .and_then(|o| o.vex.as_ref());
-
-                        if let Some(vex) = vex_json {
-                            if !vex.is_null() {
-                                let vex_bytes = serde_json::to_vec(vex)?;
-                                let mut labels = Labels::default();
-                                labels
-                                    .0
-                                    .insert("source".to_string(), "exploit-intelligence".to_string());
-
-                                let tx = db_rw.begin().await?;
-                                let result = ingestor
-                                    .ingest(
-                                        &vex_bytes,
-                                        Format::Advisory,
-                                        labels,
-                                        None,
-                                        Cache::Skip,
-                                        &tx,
-                                    )
-                                    .await
-                                    .map_err(|e| anyhow::anyhow!("VEX ingestion failed: {e}"))?;
-                                tx.commit().await?;
-
-                                tracing::info!(
-                                    job_id = %job_id,
-                                    component = image_name,
-                                    advisory_id = %result.id,
-                                    "ingested VEX advisory for SPDX component"
-                                );
-
-                                result
-                                    .id
-                                    .strip_prefix("urn:uuid:")
-                                    .and_then(|id| Uuid::from_str(id).ok())
-                            } else {
-                                None
-                            }
-                        } else {
-                            None
-                        }
-                    }
-                    _ => {
-                        tracing::warn!(
-                            job_id = %job_id,
-                            scan_id = %scan_id,
-                            "failed to fetch full report for VEX ingestion, skipping"
-                        );
-                        None
-                    }
-                };
-
-                ei_service
-                    .update_component_completed(
-                        component_id,
-                        finding,
-                        advisory_id,
-                        db_rw,
-                    )
-                    .await
-                    .map_err(|e| anyhow::anyhow!("failed to update component: {e}"))?;
-            }
-
-            // Mark the parent job as completed. The aggregate finding is
-            // computed from component records at query time.
-            ei_service
-                .update_job_completed(job_id, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-
-            tracing::info!(
-                job_id = %job_id,
-                product_id = %product_id,
-                "SPDX product analysis completed"
-            );
-
-            Ok(())
-        }
-        "failed" => {
-            let msg = format!("EI product analysis failed for product {product_id}");
-            ei_service
-                .update_job_failed(job_id, &msg, db_rw)
-                .await
-                .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
-            Ok(())
-        }
-        _ => unreachable!("backon retry only returns terminal states"),
     }
 }
