+++ modules/fundamental/src/exploit_intelligence/mod.rs
@@ -3,4 +3,5 @@ pub mod ei_client;
 pub(crate) mod endpoints;
 
 pub mod model;
+pub(crate) mod runner;
 pub mod service;
