+++ modules/fundamental/src/exploit_intelligence/runner/analysis.rs
@@ -1,0 +1,230 @@
+use crate::exploit_intelligence::{
+    runner::{EiRequestError, classify_response_error},
+    service::ExploitIntelligenceService,
+};
+use backon::{ExponentialBuilder, Retryable};
+use futures_util::TryStreamExt;
+use sea_orm::{AccessMode, ColumnTrait, EntityTrait, QueryFilter, TransactionTrait};
+use std::str::FromStr;
+use trustify_common::db;
+use trustify_module_ingestor::service::IngestorService;
+use trustify_module_storage::service::StorageBackend;
+use uuid::Uuid;
+
+use super::polling::{poll_for_product_result, poll_for_result};
+
+/// Background task: submit SBOM to EI client, poll for results, ingest VEX.
+pub(crate) async fn run_analysis(
+    ei_service: &ExploitIntelligenceService,
+    ingestor: &IngestorService,
+    db_rw: &db::ReadWrite,
+    config: &crate::exploit_intelligence::service::ExploitIntelligenceConfig,
+    client: &reqwest::Client,
+    job_id: Uuid,
+    sbom_id: Uuid,
+    vulnerability_id: &str,
+) -> Result<(), anyhow::Error> {
+    // Retrieve the raw SBOM document from storage (read-only – no mutations needed).
+    let tx = db_rw
+        .begin_with_config(None, Some(AccessMode::ReadOnly))
+        .await?;
+    let sbom_model = trustify_entity::sbom::Entity::find()
+        .filter(trustify_entity::sbom::Column::SbomId.eq(sbom_id))
+        .one(&tx)
+        .await?
+        .ok_or_else(|| anyhow::anyhow!("SBOM not found: {sbom_id}"))?;
+
+    let source_doc =
+        trustify_entity::source_document::Entity::find_by_id(sbom_model.source_document_id)
+            .one(&tx)
+            .await?
+            .ok_or_else(|| anyhow::anyhow!("source document not found for SBOM: {sbom_id}"))?;
+
+    // Build storage key from source document SHA256.
+    // The DB stores the raw hex digest without algorithm prefix, but Id::from_str expects "sha256:…".
+    let sha256_with_prefix = format!("sha256:{}", source_doc.sha256);
+    let storage_key = trustify_common::id::Id::from_str(&sha256_with_prefix)
+        .map_err(|e| anyhow::anyhow!("invalid sha256: {e}"))?;
+    let storage_key: trustify_module_storage::service::StorageKey = storage_key
+        .try_into()
+        .map_err(|e: trustify_module_storage::service::StorageKeyError| {
+            anyhow::anyhow!("storage key error: {e}")
+        })?;
+
+    let stream = ingestor
+        .storage()
+        .retrieve(storage_key)
+        .await
+        .map_err(|e| anyhow::anyhow!("storage retrieve error: {e}"))?
+        .ok_or_else(|| anyhow::anyhow!("SBOM document not found in storage"))?;
+
+    // Collect the stream into bytes.
+    let sbom_bytes: Vec<u8> = stream
+        .try_fold(Vec::new(), |mut acc, chunk| async move {
+            acc.extend_from_slice(&chunk);
+            Ok(acc)
+        })
+        .await
+        .map_err(|e| anyhow::anyhow!("failed to read SBOM stream: {e}"))?;
+
+    drop(tx);
+
+    // Detect SBOM format from labels.
+    let sbom_format = match sbom_model.labels.0.get("type") {
+        Some(t) => t.as_str(),
+        None => {
+            tracing::warn!(
+                sbom_id = %sbom_id,
+                "SBOM has no 'type' label, defaulting to CycloneDX format"
+            );
+            "cyclonedx"
+        }
+    };
+
+    let is_spdx = sbom_format == "spdx";
+
+    if is_spdx {
+        // SPDX multi-component flow: upload to /api/v1/products/upload-spdx
+        let upload_url = format!("{}/api/v1/products/upload-spdx", config.url);
+
+        let response = (|| {
+            let url = upload_url.clone();
+            let sbom = sbom_bytes.clone();
+            let vuln_id = vulnerability_id.to_string();
+            let cli = client.clone();
+            let tp = config.token_provider.clone();
+            async move {
+                let form = reqwest::multipart::Form::new()
+                    .text("cveId", vuln_id)
+                    .part(
+                        "file",
+                        reqwest::multipart::Part::bytes(sbom)
+                            .file_name("sbom.json")
+                            .mime_str("application/json")
+                            .map_err(|e| {
+                                EiRequestError::Permanent(anyhow::anyhow!("mime error: {e}"))
+                            })?,
+                    );
+
+                let mut request = cli.post(&url).multipart(form);
+
+                if let Some(ref provider) = tp {
+                    let token = provider.get_token().await.map_err(|e| {
+                        EiRequestError::Permanent(anyhow::anyhow!(
+                            "failed to obtain auth token: {e}"
+                        ))
+                    })?;
+                    request = request.bearer_auth(&token);
+                }
+
+                classify_response_error(request.send().await)
+            }
+        })
+        .retry(ExponentialBuilder::default().with_min_delay(std::time::Duration::from_secs(1)).with_max_times(3))
+        .when(|e| e.is_transient())
+        .notify(|e, dur| tracing::warn!(?dur, %e, "retrying EI upload"))
+        .await
+        .map_err(EiRequestError::into_anyhow)?;
+
+        let spdx_upload: crate::exploit_intelligence::ei_client::SpdxUploadResponse =
+            response.json().await?;
+        let product_id = spdx_upload.product_id;
+
+        // Update job to running with the product ID.
+        ei_service
+            .update_job_running_with_product(job_id, &product_id, db_rw)
+            .await
+            .map_err(|e| anyhow::anyhow!("failed to update job: {e}"))?;
+
+        tracing::info!(
+            job_id = %job_id,
+            product_id = %product_id,
+            "SPDX product uploaded to EI, starting product-level polling"
+        );
+
+        // Hand off to the product-level poll loop.
+        poll_for_product_result(
+            ei_service,
+            ingestor,
+            db_rw,
+            config,
+            client,
+            job_id,
+            vulnerability_id,
+            &product_id,
+        )
+        .await
+    } else {
+        // CycloneDX single-component flow.
+        // Create the component record before uploading so it is tracked.
+        let component_ref = sbom_model.document_id.clone().unwrap_or_else(|| sbom_model.node_id.clone());
+        let component_id = ei_service
+            .create_component(job_id, &component_ref, db_rw)
+            .await
+            .map_err(|e| anyhow::anyhow!("failed to create component: {e}"))?;
+
+        let upload_url = format!("{}/api/v1/products/upload-cyclonedx", config.url);
+
+        let response = (|| {
+            let url = upload_url.clone();
+            let sbom = sbom_bytes.clone();
+            let vuln_id = vulnerability_id.to_string();
+            let cli = client.clone();
+            let tp = config.token_provider.clone();
+            async move {
+                let form = reqwest::multipart::Form::new()
+                    .text("cveId", vuln_id)
+                    .part(
+                        "file",
+                        reqwest::multipart::Part::bytes(sbom)
+                            .file_name("sbom.json")
+                            .mime_str("application/json")
+                            .map_err(|e| {
+                                EiRequestError::Permanent(anyhow::anyhow!("mime error: {e}"))
+                            })?,
+                    );
+
+                let mut request = cli.post(&url).multipart(form);
+
+                if let Some(ref provider) = tp {
+                    let token = provider.get_token().await.map_err(|e| {
+                        EiRequestError::Permanent(anyhow::anyhow!(
+                            "failed to obtain auth token: {e}"
+                        ))
+                    })?;
+                    request = request.bearer_auth(&token);
+                }
+
+                classify_response_error(request.send().await)
+            }
+        })
+        .retry(ExponentialBuilder::default().with_min_delay(std::time::Duration::from_secs(1)).with_max_times(3))
+        .when(|e| e.is_transient())
+        .notify(|e, dur| tracing::warn!(?dur, %e, "retrying EI upload"))
+        .await
+        .map_err(EiRequestError::into_anyhow)?;
+
+        let upload: crate::exploit_intelligence::ei_client::UploadResponse = response.json().await?;
+        let scan_id = upload.report_request_id.report_id;
+
+        // Update component to running with scan ID.
+        ei_service
+            .update_component_running(component_id, &scan_id, db_rw)
+            .await
+            .map_err(|e| anyhow::anyhow!("failed to update component: {e}"))?;
+
+        // Hand off to the poll loop.
+        poll_for_result(
+            ei_service,
+            ingestor,
+            db_rw,
+            config,
+            client,
+            job_id,
+            vulnerability_id,
+            &scan_id,
+            component_id,
+        )
+        .await
+    }
+}
