+++ modules/fundamental/src/exploit_intelligence/runner/mod.rs
@@ -1,0 +1,63 @@
+pub(crate) mod analysis;
+pub(crate) mod polling;
+pub(crate) mod recovery;
+
+/// Error classification for EI HTTP requests.
+pub(crate) enum EiRequestError {
+    /// Retryable failure (network timeout, 5xx, etc.)
+    Transient(String),
+    /// Fatal failure (auth error, 4xx, parse error)
+    Permanent(anyhow::Error),
+}
+
+impl EiRequestError {
+    pub(crate) fn is_transient(&self) -> bool {
+        matches!(self, Self::Transient(_))
+    }
+
+    pub(crate) fn into_anyhow(self) -> anyhow::Error {
+        match self {
+            Self::Transient(msg) => anyhow::anyhow!(msg),
+            Self::Permanent(err) => err,
+        }
+    }
+}
+
+impl std::fmt::Display for EiRequestError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Transient(msg) => write!(f, "transient: {msg}"),
+            Self::Permanent(err) => write!(f, "permanent: {err}"),
+        }
+    }
+}
+
+/// Classify an HTTP response or reqwest error into an EiRequestError.
+pub(crate) fn classify_response_error(
+    result: Result<reqwest::Response, reqwest::Error>,
+) -> Result<reqwest::Response, EiRequestError> {
+    match result {
+        Ok(resp) => {
+            let status = resp.status();
+            if status.is_success() {
+                Ok(resp)
+            } else {
+                match status.as_u16() {
+                    408 | 429 | 502 | 503 | 504 => Err(EiRequestError::Transient(format!(
+                        "transient HTTP status {status}"
+                    ))),
+                    _ => Err(EiRequestError::Permanent(anyhow::anyhow!("HTTP {status}"))),
+                }
+            }
+        }
+        Err(err) => {
+            if err.is_connect() || err.is_timeout() || err.is_request() {
+                Err(EiRequestError::Transient(format!("network error: {err}")))
+            } else {
+                Err(EiRequestError::Permanent(anyhow::anyhow!(
+                    "network error: {err}"
+                )))
+            }
+        }
+    }
+}
