+++ modules/fundamental/src/exploit_intelligence/runner/recovery.rs
@@ -1,0 +1,116 @@
+use crate::exploit_intelligence::service::ExploitIntelligenceService;
+use trustify_common::db;
+use trustify_module_ingestor::service::IngestorService;
+
+use super::polling::{poll_for_product_result, poll_for_result};
+
+/// Recover orphaned EI jobs after a server restart.
+///
+/// Pending jobs (never submitted) are marked as Failed.  Running jobs
+/// (scan_id present) get their poll loops re-spawned so the EI backend
+/// result is picked up transparently.
+pub(crate) async fn recover_and_resume(
+    ei_service: &ExploitIntelligenceService,
+    ingestor: &IngestorService,
+    db_rw: &db::ReadWrite,
+) {
+    let result = match ei_service.recover_orphaned_jobs(db_rw).await {
+        Ok(r) => r,
+        Err(e) => {
+            tracing::error!(error = %e, "failed to recover orphaned exploit intelligence jobs");
+            return;
+        }
+    };
+
+    let runtime = match ei_service.runtime().cloned() {
+        Some(rt) => rt,
+        None => return,
+    };
+    let config = runtime.config;
+    let client = runtime.client;
+
+    for job in result.running_jobs {
+        let job_id = job.id;
+        let vulnerability_id = job.vulnerability_id.clone();
+        let ei_svc = ei_service.clone();
+        let ingestor_svc = ingestor.clone();
+        let db_rw_clone = db_rw.clone();
+        let config_clone = config.clone();
+        let client_clone = client.clone();
+
+        if let Some(ref product_id) = job.product_id {
+            // Multi-component SPDX flow — resume product-level polling.
+            let product_id = product_id.clone();
+            actix_web::rt::spawn(async move {
+                tracing::info!(job_id = %job_id, product_id = %product_id, "resuming orphaned SPDX product polling");
+                if let Err(e) = poll_for_product_result(
+                    &ei_svc,
+                    &ingestor_svc,
+                    &db_rw_clone,
+                    &config_clone,
+                    &client_clone,
+                    job_id,
+                    &vulnerability_id,
+                    &product_id,
+                )
+                .await
+                {
+                    tracing::error!(job_id = %job_id, error = %e, "failed to resume orphaned SPDX product polling");
+                    let _ = ei_svc
+                        .update_job_failed(job_id, &e.to_string(), &db_rw_clone)
+                        .await;
+                }
+            });
+        } else {
+            // Single-component CycloneDX flow — look up scan_id from the component.
+            let components = match ei_service.fetch_components(job_id, db_rw).await {
+                Ok(c) => c,
+                Err(e) => {
+                    tracing::error!(job_id = %job_id, error = %e, "failed to fetch components during recovery");
+                    let _ = ei_service
+                        .update_job_failed(job_id, &format!("recovery error: {e}"), db_rw)
+                        .await;
+                    continue;
+                }
+            };
+
+            let first_component = components.first();
+            let scan_id = first_component.and_then(|c| c.scan_id.as_deref());
+            let component_id = first_component.map(|c| c.id);
+
+            if let (Some(scan_id), Some(component_id)) = (scan_id, component_id) {
+                let scan_id = scan_id.to_string();
+                actix_web::rt::spawn(async move {
+                    tracing::info!(job_id = %job_id, scan_id = %scan_id, "resuming orphaned EI analysis");
+                    if let Err(e) = poll_for_result(
+                        &ei_svc,
+                        &ingestor_svc,
+                        &db_rw_clone,
+                        &config_clone,
+                        &client_clone,
+                        job_id,
+                        &vulnerability_id,
+                        &scan_id,
+                        component_id,
+                    )
+                    .await
+                    {
+                        tracing::error!(job_id = %job_id, error = %e, "failed to resume orphaned EI analysis");
+                        let _ = ei_svc
+                            .update_job_failed(job_id, &e.to_string(), &db_rw_clone)
+                            .await;
+                    }
+                });
+            } else {
+                tracing::error!(job_id = %job_id, "running job has no component with scan_id, marking failed");
+                let _ = ei_service
+                    .update_job_failed(
+                        job_id,
+                        "Running job missing component scan_id during recovery",
+                        db_rw,
+                    )
+                    .await;
+            }
+        }
+    }
+}
